Moves PerKeyCombineFnRunners to Flink runner
Flink is its only user. This removes the only remaining
mentions of OldDoFn in the SDK that are not OldDoFn itself.
